+++ Cmdline/Action/KSP.cs
@@ -125,6 +125,7 @@ namespace CKAN.CmdLine
             [ValueOption(1)] public string path { get; set; }
             [ValueOption(2)] public string version { get; set; }
             [ValueOption(3)] public string dlcVersion { get; set; }
+            [Option("set-default", DefaultValue = false, HelpText = "Set the new instance to the default one.")] public bool setToDefault { get; set; }
         }
 
         // This is required by ISubCommand
@@ -294,13 +295,19 @@ namespace CKAN.CmdLine
 
         private int CloneInstall(CloneOptions options)
         {
-            log.Info("Cloning the KSP instance: " + options.nameOrPath);
+            if (options.nameOrPath == null || options.new_name == null || options.new_path == null)
+            {
+                User.RaiseMessage("ksp clone <nameOrPathExistingInstance> <newName> <newPath> - argument(s) missing");
+                return Exit.BADOPT;
+            }
 
             // Parse all options
             string instanceNameOrPath = options.nameOrPath;
             string new_name = options.new_name;
             string new_path = options.new_path;
 
+
+            log.Info("Cloning the KSP instance: " + options.nameOrPath);
 
             try
             {
@@ -341,7 +348,6 @@ namespace CKAN.CmdLine
                 // Second: Something went wrong adding the new instance to the registry,
                 // most likely because the newly created directory is not valid.
 
-                User.RaiseError(kraken.ToString());
                 log.Error(kraken);
                 return Exit.ERROR;
             }
@@ -349,15 +355,13 @@ namespace CKAN.CmdLine
             {
                 // The new path is not empty
                 // The exception contains a message to inform the user.
-
-                User.RaiseError(e.ToString());
                 log.Error(e);
                 return Exit.ERROR;
             }
-            catch (NoGameInstanceKraken kraken)
+            catch (NoGameInstanceKraken)
             {
-                log.Error(kraken);
-                User.RaiseError(String.Concat(kraken.ToString(), "\n", "No instance with this name or at this path: ", instanceNameOrPath));
+                User.RaiseError(String.Format("No instance with this name or at this path: {0}\n See below for a list of known instances:\n", instanceNameOrPath));
+                ListInstalls();
                 return Exit.ERROR;
             }
 
@@ -496,52 +500,72 @@ namespace CKAN.CmdLine
         /// </summary>
         private int FakeNewKSPInstall (FakeOptions options)
         {
-            log.Info("Creating a new fake KSP install");
+            if (options.name == null || options.path == null || options.version == null)
+            {
+                User.RaiseMessage("ksp fake <name> <path> <version> [dlcVersion] - argument(s) missing");
+                return Exit.BADOPT;
+            }
 
+            log.Debug("Parsing arguments...");
             // Parse all options
             string installName = options.name;
             string path = options.path;
             KspVersion version;
-            // dlcVersion is null if no user wants no simulated DLC
-            string dlcVersion = options.dlcVersion.ToLower() == "none" ? null : options.dlcVersion;
+            bool setToDefault = options.setToDefault;
+            // dlcVersion is null if a user wants no simulated DLC
+            string dlcVersion;
+            if (options.dlcVersion == null || options.dlcVersion.ToLower() == "none")
+            {
+                dlcVersion = null;
+            }
+            else
+            {
+                dlcVersion = options.dlcVersion;
+            }
+
+            // Parse the choosen KSP version
             try
             {
-                log.Debug("Parsing KSP version");
                 version = KspVersion.Parse(options.version);
-                if (!version.IsBuildDefined)
-                {
-                    version = version.FindKnownVersion();
-                }
-                if (version == null)
-                {
-                    throw new ArgumentOutOfRangeException(nameof(version), "Your specified version is invalid.");
-                }
             }
-            catch (ArgumentOutOfRangeException e)
+            catch (FormatException)
             {
-                log.Error(e);
-                User.RaiseError(e.ToString());
-                User.RaiseMessage("Examples for valid version formats: 1.5.0 | 1.5.1 | 1.6.0.2395");
+                // Thrown if there is anything besides numbers and points in the version string or a different syntactic error.
+                User.RaiseError("Please check the version argument - Format it like Maj.Min.Patch[.Build] - e.g. 1.6.0 or 1.2.2.1622");
+                return Exit.BADOPT;
+            }
+
+            // Get the full version including build number.
+            try
+            {
+                version = version.RaiseVersionSelectionDialog(User);
+            }
+            catch (IncorrectKSPVersionKraken)
+            {
+                User.RaiseError("Couldn't find a valid KSP version for your input.\n" +
+                	"Make sure to enter the at least the version major and minor values in the form Maj.Min - e.g. 1.5");
+                return Exit.BADOPT;
+            }
+            catch (CancelledActionKraken)
+            {
+                User.RaiseError("Selection cancelled! Please call 'ckan ksp fake' again.");
                 return Exit.ERROR;
             }
+
+
+            User.RaiseMessage(String.Format("Creating new fake KSP install {0} at {1} with version {2}", installName, path, version.ToString()));
 
             try
             {
                 // Pass all arguments to CKAN.KSPManager.FakeInstance() and create a new one.
                 Manager.FakeInstance(installName, path, version, dlcVersion);
-            }
-            catch (ArgumentOutOfRangeException e)
-            {
-                // The version could not be found in the known builds.
-                log.Error(e);
-                User.RaiseError(e.ToString());
-                return Exit.ERROR;
+                if (setToDefault)
+                    User.RaiseMessage("Setting new instance to default...");
+                    Manager.SetAutoStart(installName);
             }
             catch (BadInstallLocationKraken kraken)
             {
                 // The folder exists and is not empty.
-
-                User.RaiseError(kraken.ToString());
                 log.Error(kraken);
                 return Exit.ERROR;
             }
@@ -549,10 +573,13 @@ namespace CKAN.CmdLine
             {
                 // Something went wrong adding the new instance to the registry,
                 // most likely because the newly created directory is not valid.
-
-                User.RaiseError(kraken.ToString());
                 log.Error(kraken);
                 return Exit.ERROR;
+            }
+            catch (InvalidKSPInstanceKraken)
+            {
+                // Thrown by Manager.SetAutoStart() if Manager.HasInstance returns false.
+                // Will be checked again down below with a proper error message
             }
 
 
@@ -560,6 +587,7 @@ namespace CKAN.CmdLine
             // No need to test if valid, because this is done in AddInstance().
             if (Manager.HasInstance(installName))
             {
+                User.RaiseMessage("--Done--");
                 return Exit.OK;
             }
             else